+++ chatlas/types/openai/_submit.py
@@ -38,6 +38,16 @@ class SubmitInputArgs(TypedDict, total=False):
     model: Union[
         str,
         Literal[
+            "gpt-4.1",
+            "gpt-4.1-mini",
+            "gpt-4.1-nano",
+            "gpt-4.1-2025-04-14",
+            "gpt-4.1-mini-2025-04-14",
+            "gpt-4.1-nano-2025-04-14",
+            "o4-mini",
+            "o4-mini-2025-04-16",
+            "o3",
+            "o3-2025-04-16",
             "o3-mini",
             "o3-mini-2025-01-31",
             "o1",
@@ -53,6 +63,7 @@ class SubmitInputArgs(TypedDict, total=False):
             "gpt-4o-audio-preview",
             "gpt-4o-audio-preview-2024-10-01",
             "gpt-4o-audio-preview-2024-12-17",
+            "gpt-4o-audio-preview-2025-06-03",
             "gpt-4o-mini-audio-preview",
             "gpt-4o-mini-audio-preview-2024-12-17",
             "gpt-4o-search-preview",
@@ -60,6 +71,7 @@ class SubmitInputArgs(TypedDict, total=False):
             "gpt-4o-search-preview-2025-03-11",
             "gpt-4o-mini-search-preview-2025-03-11",
             "chatgpt-4o-latest",
+            "codex-mini-latest",
             "gpt-4o-mini",
             "gpt-4o-mini-2024-07-18",
             "gpt-4-turbo",
@@ -119,7 +131,7 @@ class SubmitInputArgs(TypedDict, total=False):
         openai.NotGiven,
     ]
     seed: Union[int, None, openai.NotGiven]
-    service_tier: Union[Literal["auto", "default"], None, openai.NotGiven]
+    service_tier: Union[Literal["auto", "default", "flex"], None, openai.NotGiven]
     stop: Union[str, None, list[str], openai.NotGiven]
     store: Union[bool, None, openai.NotGiven]
     stream: Union[Literal[False], None, Literal[True], openai.NotGiven]