+++ CHANGELOG.md
@@ -7,24 +7,65 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -->
 
-## [UNRELEASED]
+## [0.8.1] - 2025-05-30
+
+* Fixed `@overload` definitions for `.stream()` and `.stream_async()`.
+
+## [0.8.0] - 2025-05-30
 
 ### New features
 
-* `.stream()` and `.stream_async()` gain a `content` argument. Set this to `"all"` to include `ContentToolRequest` and `ContentToolResponse` instances in the stream. (#75)
-* `ContentToolRequest` and `ContentToolResponse` are now exported to `chatlas` namespace. (#75)
-* `ContentToolRequest` and `ContentToolResponse` now have `.tagify()` methods, making it so they can render automatically in a Shiny chatbot. (#75)
-* `ContentToolResult` instances can be returned from tools. This allows for custom rendering of the tool result. (#75)
+* New `.on_tool_request()` and `.on_tool_result()` methods register callbacks that fire when a tool is requested or produces a result. These callbacks can be used to implement custom logging or other actions when tools are called, without modifying the tool function (#101).
+* New `ToolRejectError` exception can be thrown from tool request/result callbacks or from within a tool function itself to prevent the tool from executing. Moreover, this exception will provide some context for the the LLM to know that the tool didn't produce a result because it was rejected. (#101)
+
+### Improvements
+
+* The `CHATLAS_LOG` environment variable now enables logs for the relevant model provider. It now also supports a level of `debug` in addition to `info`. (#97)
+* `ChatSnowflake()` now supports tool calling. (#98)
+* `Chat` instances can now be deep copied, which is useful for forking the chat session. (#96)
+
+### Changes
+
+* `ChatDatabricks()`'s `model` now defaults to `databricks-claude-3-7-sonnet` instead of `databricks-dbrx-instruct`. (#95)
+* `ChatSnowflake()`'s `model` now defaults to `claude-3-7-sonnet` instead of `llama3.1-70b`. (#98)
+
+### Bug fixes
+
+* Fixed an issue where `ChatDatabricks()` with an Anthropic `model` wasn't handling empty-string responses gracefully. (#95)
+
+
+## [0.7.1] - 2025-05-10
+
+* Added `openai` as a hard dependency, making installation easier for a wide range of use cases. (#91) 
+
+## [0.7.0] - 2025-04-22
+
+### New features
+
+* Added `ChatDatabricks()`, for chatting with Databrick's [foundation models](https://docs.databricks.com/aws/en/machine-learning/model-serving/score-foundation-models). (#82)
+* `.stream()` and `.stream_async()` gain a `content` argument. Set this to `"all"` to include `ContentToolResult`/`ContentToolRequest` objects in the stream. (#75)
+* `ContentToolResult`/`ContentToolRequest` are now exported to `chatlas` namespace. (#75)
+* `ContentToolResult`/`ContentToolRequest` gain a `.tagify()` method so they render sensibly in a Shiny app. (#75)
+* A tool can now return a `ContentToolResult`. This is useful for: 
+    * Specifying the format used for sending the tool result to the chat model (`model_format`). (#87)
+    * Custom rendering of the tool result (by overriding relevant methods in a subclass). (#75)
 * `Chat` gains a new `.current_display` property. When a `.chat()` or `.stream()` is currently active, this property returns an object with a `.echo()` method (to echo new content to the display). This is primarily useful for displaying custom content during a tool call. (#79)
 * Added support for `Chat` providers to be [MCP clients](https://modelcontextprotocol.io/) and register with remote tools over SSE or stdio.
 
 ### Improvements
 
 * When a tool call ends in failure, a warning is now raised and the stacktrace is printed. (#79)
+* Several improvements to `ChatSnowflake()`:
+  * `.extract_data()` is now supported.
+  *  `async` methods are now supported. (#81)
+  * Fixed an issue with more than one session being active at once. (#83)
+* `ChatAnthropic()` no longer chokes after receiving an output that consists only of whitespace. (#86)
+* `orjson` is now used for JSON loading and dumping. (#87)
 
 ### Changes
 
 * The `echo` argument of the `.chat()` method defaults to a new value of `"output"`. As a result, tool requests and results are now echoed by default. To revert to the previous behavior, set `echo="text"`. (#78)
+* Tool results are now dumped to JSON by default before being sent to the model. To revert to the previous behavior, have the tool return a `ContentToolResult` with `model_format="str"`. (#87)
 
 ### Breaking changes